refactor: update volume store
