+++ internal/adapter/store/volume/configmap.go
@@ -3,7 +3,6 @@ package volume
 import (
 	"context"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/docker/docker/api/types/volume"
@@ -15,11 +14,19 @@ import (
 	"k8s.io/kubernetes/pkg/apis/core"
 )
 
-// TODO: update function comments with namespace support
-
-// DeleteConfigMap removes a Docker volume corresponding to a Kubernetes ConfigMap.
-// The volume name is generated by appending a predefined prefix to the ConfigMap name.
-// Returns an error if the Docker volume could not be removed; otherwise, returns nil.
+// DeleteConfigMap deletes a specific ConfigMap identified by its name and namespace
+// from a Docker volume-based ConfigMap store.
+//
+// The function performs the following steps:
+// 1. Builds the Docker volume name for the ConfigMap based on its name and namespace.
+// 2. Removes the Docker volume associated with the ConfigMap.
+//
+// Parameters:
+// - configMapName: The name of the ConfigMap to delete.
+// - namespace: The namespace where the ConfigMap is located.
+//
+// Returns:
+// - An error object if the function fails to delete the ConfigMap.
 func (store *VolumeStore) DeleteConfigMap(configMapName, namespace string) error {
 	volumeName := buildConfigMapVolumeName(configMapName, namespace)
 
@@ -40,9 +47,22 @@ func (s *VolumeStore) GetConfigMapBinds(configMap *core.ConfigMap) (map[string]s
 	}, nil
 }
 
-// GetConfigMap fetches a Kubernetes ConfigMap represented as a Docker volume.
-// The function performs a Docker volume inspect operation and converts it into a ConfigMap.
-// Returns a pointer to a ConfigMap object, and an error if any occurs.
+// GetConfigMap retrieves a specific ConfigMap identified by its name and namespace
+// from a Docker volume-based ConfigMap store.
+//
+// The function performs the following steps:
+// 1. Builds the Docker volume name for the ConfigMap based on its name and namespace.
+// 2. Inspects the Docker volume to retrieve its details.
+// 3. Creates a ConfigMap object from the inspected Docker volume.
+// 4. Fetches the data map from the Docker volume and associates it with the ConfigMap object.
+//
+// Parameters:
+// - configMapName: The name of the ConfigMap to retrieve.
+// - namespace: The namespace where the ConfigMap is located.
+//
+// Returns:
+// - A pointer to the retrieved ConfigMap object.
+// - An error object if the function fails to retrieve the ConfigMap.
 func (s *VolumeStore) GetConfigMap(configMapName, namespace string) (*core.ConfigMap, error) {
 	volumeName := buildConfigMapVolumeName(configMapName, namespace)
 
@@ -54,7 +74,7 @@ func (s *VolumeStore) GetConfigMap(configMapName, namespace string) (*core.Confi
 		return nil, fmt.Errorf("unable to inspect Docker volume: %w", err)
 	}
 
-	configMap, err := buildConfigMapFromVolume(&volume)
+	configMap, err := createConfigMapFromVolume(&volume)
 	if err != nil {
 		return nil, fmt.Errorf("unable to build config map from volume: %w", err)
 	}
@@ -69,9 +89,21 @@ func (s *VolumeStore) GetConfigMap(configMapName, namespace string) (*core.Confi
 	return &configMap, nil
 }
 
-// GetConfigMaps lists all Kubernetes ConfigMaps represented as Docker volumes.
-// Filters are applied to list only the volumes that are meant to be ConfigMaps.
-// Returns a ConfigMapList object containing all matched ConfigMaps, and an error if any occurs.
+// GetConfigMaps retrieves all ConfigMaps for a given namespace from a
+// Docker volume-based ConfigMap store.
+//
+// The function performs the following steps:
+// 1. Creates a filter to list Docker volumes associated with ConfigMaps in the given namespace.
+// 2. Lists the Docker volumes using the created filter.
+// 3. Creates ConfigMap objects from the listed Docker volumes.
+// 4. Fetches data maps from the Docker volumes and associates them with the ConfigMap objects.
+//
+// Parameters:
+// - namespace: The namespace for which to retrieve ConfigMaps.
+//
+// Returns:
+// - A ConfigMapList object containing all the ConfigMaps for the given namespace.
+// - An error object if the function fails to retrieve the ConfigMaps.
 func (store *VolumeStore) GetConfigMaps(namespace string) (core.ConfigMapList, error) {
 	filter := configMapListFilter(namespace)
 	volumes, err := store.cli.VolumeList(context.TODO(), volume.ListOptions{
@@ -100,7 +132,7 @@ func (store *VolumeStore) GetConfigMaps(namespace string) (core.ConfigMapList, e
 	}
 
 	for _, volume := range volumes.Volumes {
-		configMap, err := buildConfigMapFromVolume(volume)
+		configMap, err := createConfigMapFromVolume(volume)
 		if err != nil {
 			store.logger.Warnf("unable to build config map from volume %s: %w", volume.Name, err)
 			continue
@@ -114,9 +146,18 @@ func (store *VolumeStore) GetConfigMaps(namespace string) (core.ConfigMapList, e
 	return configMaps, nil
 }
 
-// StoreConfigMap stores a Kubernetes ConfigMap as a Docker volume.
-// The volume is labeled to indicate that it represents a ConfigMap.
-// Returns an error if the Docker volume could not be created; otherwise, returns nil.
+// StoreConfigMap stores a given ConfigMap object in a Docker volume-based ConfigMap store.
+//
+// The function performs the following steps:
+// 1. Builds the Docker volume name for the ConfigMap based on its name and namespace.
+// 2. Creates a new Docker volume with the constructed name and attaches labels to it.
+// 3. Copies the data map of the ConfigMap to the created Docker volume.
+//
+// Parameters:
+// - configMap: A pointer to the ConfigMap object to store.
+//
+// Returns:
+// - An error object if the function fails to store the ConfigMap.
 func (store *VolumeStore) StoreConfigMap(configMap *corev1.ConfigMap) error {
 	volumeName := buildConfigMapVolumeName(configMap.Name, configMap.Namespace)
 
@@ -142,19 +183,9 @@ func (store *VolumeStore) StoreConfigMap(configMap *corev1.ConfigMap) error {
 	return nil
 }
 
-// Each configmap is stored as a Docker volume using the following naming convention:
-// configmap-[namespace]-[configmap-name]
-func buildConfigMapVolumeName(configMapName, namespace string) string {
-	return fmt.Sprintf("%s%s-%s", ConfigMapVolumePrefix, namespace, configMapName)
-}
-
-func getConfigMapNameFromVolumeName(volumeName, namespace string) string {
-	return strings.TrimPrefix(volumeName, fmt.Sprintf("%s%s-", ConfigMapVolumePrefix, namespace))
-}
-
-// buildConfigMapFromVolume constructs a Kubernetes ConfigMap object from a Docker volume.
+// createConfigMapFromVolume constructs a Kubernetes ConfigMap object from a Docker volume.
 // Returns a ConfigMap object, and an error if any occurs (e.g., if the volume's creation timestamp is not parseable).
-func buildConfigMapFromVolume(volume *volume.Volume) (core.ConfigMap, error) {
+func createConfigMapFromVolume(volume *volume.Volume) (core.ConfigMap, error) {
 	namespace := volume.Labels[NamespaceNameLabelKey]
 
 	configMap := core.ConfigMap{
