+++ internal/adapter/store/volume/secret.go
@@ -3,7 +3,6 @@ package volume
 import (
 	"context"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/docker/docker/api/types/volume"
@@ -16,11 +15,19 @@ import (
 	"k8s.io/kubernetes/pkg/apis/core"
 )
 
-// TODO: update function comments with namespace support
-
-// DeleteSecret removes a Docker volume corresponding to a Kubernetes Secret.
-// The volume name is generated by appending a predefined prefix to the Secret name.
-// Returns an error if the Docker volume could not be removed; otherwise, returns nil.
+// DeleteSecret deletes a specific secret identified by its name and namespace
+// from a Docker volume-based secret store.
+//
+// The function performs the following steps:
+// 1. Builds the Docker volume name for the secret based on its name and namespace.
+// 2. Removes the Docker volume associated with the secret.
+//
+// Parameters:
+// - secretName: The name of the secret to delete.
+// - namespace: The namespace where the secret is located.
+//
+// Returns:
+// - An error object if the function fails to delete the secret.
 func (s *VolumeStore) DeleteSecret(secretName, namespace string) error {
 	volumeName := buildSecretVolumeName(secretName, namespace)
 
@@ -41,9 +48,22 @@ func (s *VolumeStore) GetSecretBinds(secret *core.Secret) (map[string]string, er
 	}, nil
 }
 
-// GetSecret fetches a Kubernetes Secret represented as a Docker volume.
-// The function performs a Docker volume inspect operation and converts it into a Secret.
-// Returns a pointer to a Secret object, and an error if any occurs.
+// GetSecret retrieves a specific secret identified by its name and namespace
+// from a Docker volume-based secret store.
+//
+// The function performs the following steps:
+// 1. Builds the Docker volume name for the secret based on its name and namespace.
+// 2. Inspects the Docker volume to retrieve its details.
+// 3. Creates a Secret object from the inspected Docker volume.
+// 4. Fetches the data map from the Docker volume and associates it with the Secret object.
+//
+// Parameters:
+// - secretName: The name of the secret to retrieve.
+// - namespace: The namespace where the secret is located.
+//
+// Returns:
+// - A pointer to the retrieved Secret object.
+// - An error object if the function fails to retrieve the secret.
 func (s *VolumeStore) GetSecret(secretName, namespace string) (*core.Secret, error) {
 	volumeName := buildSecretVolumeName(secretName, namespace)
 
@@ -55,7 +75,7 @@ func (s *VolumeStore) GetSecret(secretName, namespace string) (*core.Secret, err
 		return nil, fmt.Errorf("unable to inspect Docker volume: %w", err)
 	}
 
-	secret, err := buildSecretFromVolume(&volume)
+	secret, err := createSecretFromVolume(&volume)
 	if err != nil {
 		return nil, fmt.Errorf("unable to build secret from volume: %w", err)
 	}
@@ -70,9 +90,23 @@ func (s *VolumeStore) GetSecret(secretName, namespace string) (*core.Secret, err
 	return &secret, nil
 }
 
-// GetSecrets lists all Kubernetes Secrets represented as Docker volumes.
-// Filters are applied to list only the volumes that are meant to be Secrets.
-// Returns a SecretList object containing all matched Secrets, and an error if any occurs.
+// GetSecrets retrieves all secrets for a given namespace from a Docker volume-based secret store,
+// optionally filtered by a set of labels.
+//
+// The function performs the following steps:
+// 1. Creates a filter to list Docker volumes associated with secrets in the given namespace.
+// 2. Lists the Docker volumes using the created filter.
+// 3. Filters volumes based on label selectors.
+// 4. Creates Secret objects from the filtered Docker volumes.
+// 5. Fetches data maps from the Docker volumes and associates them with the Secret objects.
+//
+// Parameters:
+// - namespace: The namespace for which to retrieve secrets.
+// - selector: Label selector to filter secrets.
+//
+// Returns:
+// - A SecretList object containing all the filtered secrets for the given namespace.
+// - An error object if the function fails to retrieve the secrets.
 func (s *VolumeStore) GetSecrets(namespace string, selector labels.Selector) (core.SecretList, error) {
 	filter := secretListFilter(namespace)
 	volumes, err := s.cli.VolumeList(context.TODO(), volume.ListOptions{
@@ -108,7 +142,7 @@ func (s *VolumeStore) GetSecrets(namespace string, selector labels.Selector) (co
 	}
 
 	for _, volume := range filteredVolumes {
-		secret, err := buildSecretFromVolume(&volume)
+		secret, err := createSecretFromVolume(&volume)
 		if err != nil {
 			s.logger.Warnf("unable to build secret from volume %s: %w", volume.Name, err)
 			continue
@@ -122,9 +156,18 @@ func (s *VolumeStore) GetSecrets(namespace string, selector labels.Selector) (co
 	return secrets, nil
 }
 
-// StoreSecret stores a Kubernetes Secret as a Docker volume.
-// The volume is labeled to indicate that it represents a Secret.
-// Returns an error if the Docker volume could not be created; otherwise, returns nil.
+// StoreSecret stores a given Secret object in a Docker volume-based secret store.
+//
+// The function performs the following steps:
+// 1. Builds the Docker volume name for the secret based on its name and namespace.
+// 2. Creates a new Docker volume with the constructed name and attaches labels to it.
+// 3. Copies both the data map and string data of the Secret to the created Docker volume.
+//
+// Parameters:
+// - secret: A pointer to the Secret object to store.
+//
+// Returns:
+// - An error object if the function fails to store the secret.
 func (s *VolumeStore) StoreSecret(secret *corev1.Secret) error {
 	volumeName := buildSecretVolumeName(secret.Name, secret.Namespace)
 
@@ -160,19 +203,9 @@ func (s *VolumeStore) StoreSecret(secret *corev1.Secret) error {
 	return nil
 }
 
-// Each secret is stored as a Docker volume using the following naming convention:
-// secret-[namespace]-[secret-name]
-func buildSecretVolumeName(configMapName, namespace string) string {
-	return fmt.Sprintf("%s%s-%s", SecretVolumePrefix, namespace, configMapName)
-}
-
-func getSecretNameFromVolumeName(volumeName, namespace string) string {
-	return strings.TrimPrefix(volumeName, fmt.Sprintf("%s%s-", SecretVolumePrefix, namespace))
-}
-
-// buildSecretFromVolume constructs a Kubernetes Secret object from a Docker volume.
+// createSecretFromVolume constructs a Kubernetes Secret object from a Docker volume.
 // Returns a Secret object, and an error if any occurs (e.g., if the volume's creation timestamp is not parseable).
-func buildSecretFromVolume(volume *volume.Volume) (core.Secret, error) {
+func createSecretFromVolume(volume *volume.Volume) (core.Secret, error) {
 	namespace := volume.Labels[NamespaceNameLabelKey]
 
 	secret := core.Secret{
